move proptest strategies to testutils-proptest module
