+++ soroban-sdk/Cargo.toml
@@ -69,8 +69,10 @@ trybuild = "1.0.115"
 [features]
 alloc = []
 testutils = ["soroban-sdk-macros/testutils", "soroban-env-host/testutils", "soroban-ledger-snapshot/testutils", "dep:ed25519-dalek", "dep:arbitrary", "dep:derive_arbitrary", "dep:ctor", "dep:soroban-ledger-snapshot"]
-proptest = ["testutils", "dep:proptest", "dep:proptest-arbitrary-interop"]
 docs = []
+
+# Support for test frameworks and utilities
+testutils-proptest = ["testutils", "dep:proptest", "dep:proptest-arbitrary-interop"]
 
 # Umbrella feature that enables all hazmat sub-features (backwards compatible)
 hazmat = [
