+++ soroban-sdk/src/testutils/proptest.rs
@@ -1,0 +1,194 @@
+//! Support for property testing Soroban contracts with [`proptest`].
+//!
+//! This module provides the `arb` and `arb_sized` strategies, which generate
+//! the prototypes defined by the [`arbitrary`][crate::testutils::arbitrary]
+//! module, and reexports the [`proptest`] crate.
+//!
+//! This module is only available when the "testutils-proptest" Cargo feature
+//! is defined, which implies the "testutils" feature.
+//!
+//! [`proptest`]: https://github.com/proptest-rs/proptest/
+//!
+//! ```toml
+//! [dev-dependencies]
+//! proptest = "1"
+//! soroban-sdk = { version = "*", features = ["testutils-proptest"] }
+//! ```
+//!
+//! `arb::<T>()` generates the prototype of the Soroban type `T`. As with
+//! fuzzing, prototypes are converted to contract types with [`FromVal`] or
+//! [`IntoVal`], which require an [`Env`]. An `Env` cannot be carried through a
+//! strategy, so the conversion happens inside the test body.
+//!
+//! [`Env`]: crate::Env
+//! [`FromVal`]: crate::FromVal
+//! [`IntoVal`]: crate::IntoVal
+//! [`contracttype`]: crate::contracttype
+//!
+//! ```
+//! use proptest::prelude::*;
+//! use soroban_sdk::testutils::proptest::arb;
+//! use soroban_sdk::{Address, Env, IntoVal, Vec};
+//!
+//! proptest! {
+//!     #[test]
+//!     fn test_address(address in arb::<Address>()) {
+//!         let env = Env::default();
+//!         let address: Address = address.into_val(&env);
+//!         // test the contract with the address
+//!     }
+//!
+//!     #[test]
+//!     fn test_addresses(addresses in arb::<Vec<Address>>()) {
+//!         let env = Env::default();
+//!         let addresses: Vec<Address> = addresses.into_val(&env);
+//!         // test the contract with the addresses
+//!     }
+//! }
+//! ```
+//!
+//! Types with the [`contracttype`] attribute work the same way.
+//!
+//! ```
+//! use proptest::prelude::*;
+//! use soroban_sdk::testutils::proptest::arb;
+//! use soroban_sdk::{contracttype, Address, Env, IntoVal};
+//!
+//! #[contracttype]
+//! pub struct Deposit {
+//!     pub address: Address,
+//!     pub amount: i128,
+//! }
+//!
+//! proptest! {
+//!     #[test]
+//!     fn test_deposit(deposit in arb::<Deposit>()) {
+//!         let env = Env::default();
+//!         let deposit: Deposit = deposit.into_val(&env);
+//!         // test the contract with the deposit
+//!     }
+//! }
+//! ```
+//!
+//! ## Entropy budget
+//!
+//! `arb` feeds a fixed 256 bytes of random data to the `Arbitrary`
+//! implementation of the prototype. When a prototype needs more bytes than
+//! that, `arbitrary` zero-fills the remainder rather than returning an error,
+//! so the tail of a large prototype — for example the last fields of a
+//! `contracttype` struct with many `BytesN<32>` fields — is all zeros in every
+//! generated value. Use `arb_sized` with a larger budget for such types.
+//!
+//! ## Limits of generated values
+//!
+//! Values come from the [`Arbitrary`][::arbitrary::Arbitrary] implementations
+//! of the prototypes, so the limits of those implementations apply here too —
+//! collection lengths are small, and addresses are always contract addresses.
+//! See the [`arbitrary`][crate::testutils::arbitrary] module docs for the
+//! details.
+
+use super::arbitrary::SorobanArbitrary;
+use ::proptest::strategy::Strategy;
+use core::fmt::Debug;
+
+/// A reexport of the `proptest` crate.
+pub use ::proptest;
+
+/// A [`proptest`] strategy that generates prototypes of the Soroban type `T`.
+///
+/// Prototypes are converted to their contract type with [`IntoVal`] or
+/// [`FromVal`], which require an [`Env`]. Because an `Env` cannot be
+/// carried through a strategy, the conversion is performed inside the test
+/// body.
+///
+/// Values are generated by feeding random bytes to the [`Arbitrary`]
+/// implementation of the prototype, with a fixed entropy budget of 256
+/// bytes. A prototype that needs more bytes than that is not an error:
+/// the `arbitrary` crate zero-fills the remainder, so the tail of a large
+/// prototype will be all zeros in every generated value. Use
+/// [`arb_sized`] with a larger budget for large types.
+///
+/// Collection lengths are chosen by `arbitrary`, not by proptest, and so
+/// are small — see the [`arbitrary`][crate::testutils::arbitrary] module docs
+/// for the distribution and how to generate larger collections.
+///
+/// [`proptest`]: ::proptest
+/// [`Arbitrary`]: ::arbitrary::Arbitrary
+/// [`Env`]: crate::Env
+/// [`FromVal`]: crate::FromVal
+/// [`IntoVal`]: crate::IntoVal
+///
+/// # Example
+///
+/// ```
+/// use proptest::prelude::*;
+/// use soroban_sdk::testutils::proptest::arb;
+/// use soroban_sdk::{Address, Env, IntoVal};
+///
+/// proptest! {
+///     #[test]
+///     fn test_address(address in arb::<Address>()) {
+///         let env = Env::default();
+///         let address: Address = address.into_val(&env);
+///         // test the contract with the address
+///     }
+/// }
+/// ```
+pub fn arb<T>() -> impl Strategy<Value = T::Prototype>
+where
+    T: SorobanArbitrary,
+    T::Prototype: Debug + Clone + 'static,
+{
+    proptest_arbitrary_interop::arb::<T::Prototype>()
+}
+
+/// The same as [`arb`], but with a configurable entropy budget.
+///
+/// `size` is the number of random bytes fed to the [`Arbitrary`]
+/// implementation of the prototype. [`arb`] uses a fixed budget of 256
+/// bytes, which is not enough for large prototypes: when the bytes run out
+/// `arbitrary` zero-fills the rest instead of erroring, and the tail of the
+/// prototype is all zeros in every generated value. Pass a larger `size`
+/// for prototypes that need more entropy, such as a contract type with
+/// many `BytesN` fields.
+///
+/// A larger `size` also makes shrinking slower: shrinking works by
+/// truncating the byte budget one byte at a time and regenerating the
+/// value, so a failing case takes up to `size` steps to shrink, and the
+/// counterexample reported is not necessarily minimal.
+///
+/// [`Arbitrary`]: ::arbitrary::Arbitrary
+///
+/// # Example
+///
+/// ```
+/// use proptest::prelude::*;
+/// use soroban_sdk::testutils::proptest::arb_sized;
+/// use soroban_sdk::{contracttype, BytesN, Env, IntoVal};
+///
+/// // Twelve 32-byte fields need 384 bytes, more than the 256 bytes `arb`
+/// // provides, so without a larger budget the last fields are all zeros.
+/// #[contracttype]
+/// pub struct Keys {
+///     pub a: BytesN<32>, pub b: BytesN<32>, pub c: BytesN<32>,
+///     pub d: BytesN<32>, pub e: BytesN<32>, pub f: BytesN<32>,
+///     pub g: BytesN<32>, pub h: BytesN<32>, pub i: BytesN<32>,
+///     pub j: BytesN<32>, pub k: BytesN<32>, pub l: BytesN<32>,
+/// }
+///
+/// proptest! {
+///     #[test]
+///     fn test_keys(keys in arb_sized::<Keys>(1024)) {
+///         let env = Env::default();
+///         let keys: Keys = keys.into_val(&env);
+///         // test the contract with the keys
+///     }
+/// }
+/// ```
+pub fn arb_sized<T>(size: usize) -> impl Strategy<Value = T::Prototype>
+where
+    T: SorobanArbitrary,
+    T::Prototype: Debug + Clone + 'static,
+{
+    proptest_arbitrary_interop::arb_sized::<T::Prototype>(size)
+}
